[vortex] Move VectorDedicated E2E tests to paimon-vortex-format to avoid circular dependency
